+++ typescript/http-registry-server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hyperlane-xyz/http-registry-server
 
+## 19.3.0
+
+### Patch Changes
+
+- Updated dependencies [fcdda58a8]
+  - @hyperlane-xyz/sdk@19.3.0
+  - @hyperlane-xyz/utils@19.3.0
+
 ## 19.2.0
 
 ### Patch Changes