+++ typescript/radix-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hyperlane-xyz/radix-sdk
 
+## 19.3.0
+
+### Patch Changes
+
+- @hyperlane-xyz/utils@19.3.0
+
 ## 19.2.0
 
 ### Patch Changes