+++ typescript/tsconfig/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperlane-xyz/tsconfig
 
+## 19.3.0
+
 ## 19.2.0
 
 ## 19.1.1